+++ ptarmd/ptarmd.h
@@ -250,7 +250,7 @@ typedef struct {
     uint16_t    max_accepted_htlcs;                 ///< 2 : max-accepted-htlcs
     uint32_t    min_depth;                          ///< 4 : minimum-depth(acceptのみ)
 
-    uint8_t     localfeatures;                      ///< init.localfeatures
+    uint16_t    localfeatures;                      ///< init.localfeatures
     uint16_t    feerate_min;                        ///< feerate_per_kw limit min percent(0..all OK)
     uint16_t    feerate_max;                        ///< feerate_per_kw limit max percent(0..all OK)
 } channel_conf_t;
